[Fix] ネットワークからゲーム情報を取得する際、ゲームバージョンの比較にInt.Parseをかけていたため、オーバーフローが起こる不具合。long.Parseに修正

+++ Assets/Scripts/GameData/GameDatasSingleton.cs
@@ -81,7 +81,7 @@ public class GameDatasSingleton : BasedSingleton<GameDatasSingleton>
             if(singletonGameData.GameID == gameData.GameID)
             {
                 //追加されるゲームのバージョンが新しい場合
-                if(int.Parse(singletonGameData.GameVersion) < int.Parse(gameData.GameVersion) && singletonGameData.Status == GameStatus.Downloaded)
+                if(long.Parse(singletonGameData.GameVersion) < long.Parse(gameData.GameVersion) && singletonGameData.Status == GameStatus.Downloaded)
                 {
                     singletonGameData.Status = GameStatus.UpdateAvailable;
                 }
